Set level range to `int16`
This is based on querying the `v2025_10_22_0` release and finding that
the range of levels in the rest of the dataset is -9,999 to 1,940.

And really, do you need billions of levels? That's not the intention
for Z-order stacking.

+++ packages/overture-schema-core/src/overture/schema/core/types.py
@@ -4,7 +4,7 @@ from pydantic import (
     Field,
 )
 
-from overture.schema.system.primitive import float32, int32
+from overture.schema.system.primitive import float32, int16, int32
 
 ConfidenceScore = NewType(
     "ConfidenceScore",
@@ -18,7 +18,7 @@ ConfidenceScore = NewType(
 Level = NewType(
     "Level",
     Annotated[
-        int32,
+        int16,
         Field(description="Z-order of the feature where 0 is visual level"),
     ],
 )
@@ -76,7 +76,7 @@ SortKey = NewType(
     Annotated[
         int,
         Field(
-            #default=0,
+            # default=0,
             description="An ascending numeric that defines the recommended order features should be drawn in. Features with lower number should be shown on top of features with a higher number.",
         ),
     ],
